Add Docker Image and Docker Compose
Docker Image is based off a py3.12 base image. It installs a postgres
driver, an aws rds cert (via curl), and poetry. Then it copies the
pyproject toml file and the py source code and builds it.

Docker Compose file uses this docker image and a postgres15.4 image to
stand up the application for local testing.

.env.template file contains all the evn variables that need to be
defined to run locally.

update readme with more dev steps to get the docker images up and
running.