FIX: Only load initial snapshot folders from the archive bucket(#24)
ODS-QLIK table loading for EDW.PATRON_TRIP is currently not functional.

This is because the latest initial snapshot folder (based on timestamp) is currently located in the data-platform error bucket. While the snapshot folder that the cubic qlik process is populating is in the archive bucket.

Having the ods-qlik loader only look for snapshot files in the archive bucket will fix this issue for the EDW.PATRON_TRIP table. However, if an initial snapshot folder is moved to the error bucket, and remains in-use by the cubic ods-qlik process, then this functionality will again be broken.