fix: run DB migrations on staging
This change adds checks to migrations files that will allow them to run/finish on the staging evironment.

Some migrations were failing on staging because the `ods` schema has no tables on staging. This change updates the migrations that modify the ods schema to first check if tables exist in the `ods` schema. If no tables exist, the migration exits early