fix: update dmap-import schemas to match API
This change updates the schemas of a few dmap-import tables to match the cubic `schemas` endpoint for the same tables.

This change should resolve a recent error where records were not able to be loaded into the `use_transaction_location` table because of an overflow in the `serial_nbr` column.