fix: auto-convert STRING type columns for Qlik process (#68)
This change updates the ods-qlik schema verification process to automatically handle columns that were converted to STRING/VARCHAR types.

Any column type allows for conversion to VARCHAR in postgres so this change will have no negative impacts on the process.

This change will also un-stick the EDW.REASON_DIMENSION table loading and allow it to complete.